Cast FE_value to an (unsigned int) to avoid a warning generated with an implicit cast.
git-svn-id: https://svn.physiomeproject.org/svn/cmiss/cmgui/trunk@4325 3c1deb5b-d424-0410-962d-aba41a686d42

+++ source/image_processing/computed_field_region_label.c
@@ -515,7 +515,7 @@ DESCRIPTION : Implement region labeling on image cache. The image must be black-
 			        	result_index -= image->depth;
 				}
 			}
-			ALLOCATE(hist, int, label_index);
+			ALLOCATE(hist, int, (unsigned int)label_index);
 			for (k = 1; k <= label_index; k++)
 			{
 			        for ( i = 0; i < storage_size / image->depth; i++)
